fix: align difficulty_score column type, fix test compilation, respect Flyway enabled flag
- Change V13 migration from NUMERIC(4,2) to DOUBLE PRECISION to match JPA double field
- Add @ConditionalOnProperty to FlywayConfig so H2 tests skip Flyway
- Fix SoloGameControllerTest: inject Principal via .principal() on MockMvc requests
- Fix AnswerRepositoryTest: import JpaConfig for @EnableJpaAuditing
- Remove stale playerId references from FuzzyMatchingContainerTest
- Minor formatting fix in frontend page.tsx

+++ backend/src/main/resources/db/migration/V13__question_difficulty_metrics.sql
@@ -55,7 +55,7 @@ COMMENT ON COLUMN questions.viability_exclusion_reason IS
 -- ── Difficulty-score columns ───────────────────────────────────────────────
 
 ALTER TABLE questions
-    ADD COLUMN IF NOT EXISTS difficulty_score   NUMERIC(4,2) NOT NULL DEFAULT 5.00;
+    ADD COLUMN IF NOT EXISTS difficulty_score   DOUBLE PRECISION NOT NULL DEFAULT 5.0;
 COMMENT ON COLUMN questions.difficulty_score IS
     'Continuous difficulty value 0.00–10.00 computed from zone counts by DifficultyCalculator. '
     '0.00 = easiest, 10.00 = hardest. Derived labels (Easy/Hard) are computed at render time '
